requirements.txt
flask==2.3.3
werkzeug==2.3.7
pillow==10.0.1
opencv-python==4.8.1.78
pytesseract==0.3.10
PyPDF2==3.0.1
pdfplumber==0.9.0
pandas>=2.1.1
numpy>=1.26.0
scikit-learn==1.3.0
regex==2023.8.8
python-dateutil==2.8.2
openpyxl==3.1.2
jinja2==3.1.2
requests==2.31.0
cryptography==41.0.4
PyMuPDF==1.23.8
